rename rpcService to extHostRpcService
