+++ src/vs/workbench/api/common/extHost.api.impl.ts
@@ -67,7 +67,7 @@ import { IExtHostDebugService } from 'vs/workbench/api/common/extHostDebugServic
 import { IExtHostSearch } from 'vs/workbench/api/common/extHostSearch';
 import { ILogService } from 'vs/platform/log/common/log';
 import { IURITransformerService } from 'vs/workbench/api/common/extHostUriTransformerService';
-import { IExtHostRpcService } from 'vs/workbench/api/common/rpcService';
+import { IExtHostRpcService } from 'vs/workbench/api/common/extHostRpcService';
 import { IExtHostInitDataService } from 'vs/workbench/api/common/extHostInitDataService';
 
 export interface IExtensionApiFactory {
